requirements.txt
absl-py==2.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
Farama-Notifications==0.0.4
ffmpeg==1.4
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
grpcio==1.68.1
gymnasium==1.0.0
idna==3.4
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
mss==10.0.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
protobuf==5.29.1
PyGetWindow==0.0.9
pyparsing==3.2.0
PyRect==0.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.28.1
six==1.17.0
stable_baselines3==2.4.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
